refactor: move validation into internal
