+++ packages/core/src/internal/types/validation.ts
@@ -1,10 +1,6 @@
-import { CallPoints } from "../internal/types/deploymentGraph";
-import {
-  ResultsAccumulator,
-  VertexVisitResult,
-  VisitResult,
-} from "../internal/types/graph";
-import { Services } from "../internal/types/services";
+import { CallPoints } from "./deploymentGraph";
+import { ResultsAccumulator, VertexVisitResult, VisitResult } from "./graph";
+import { Services } from "./services";
 
 export type ValidationVisitResult = VisitResult<undefined>;
