+++ packages/core/src/internal/validation/dispatch/validateArtifactLibrary.ts
@@ -1,61 +1,59 @@
 import { ethers } from "ethers";
 
-import { HardhatLibraryDeploymentVertex } from "../../internal/types/deploymentGraph";
-import { VertexResultEnum } from "../../internal/types/graph";
+import { ArtifactLibraryDeploymentVertex } from "../../types/deploymentGraph";
+import { VertexResultEnum } from "../../types/graph";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
   ValidationVertexVisitResult,
 } from "../../types/validation";
+import { isArtifact } from "../../utils/guards";
 
 import { buildValidationError, validateBytesForArtifact } from "./helpers";
 
-export async function validateHardhatLibrary(
-  vertex: HardhatLibraryDeploymentVertex,
+export async function validateArtifactLibrary(
+  vertex: ArtifactLibraryDeploymentVertex,
   _resultAccumulator: ValidationResultsAccumulator,
-  { callPoints, services }: ValidationDispatchContext
+  context: ValidationDispatchContext
 ): Promise<ValidationVertexVisitResult> {
   if (!ethers.utils.isAddress(vertex.from)) {
     return buildValidationError(
       vertex,
       `For library 'from' must be a valid address string`,
-      callPoints
+      context.callPoints
     );
   }
 
   const invalidBytes = await validateBytesForArtifact({
     vertex,
-    callPoints,
-    services,
+    callPoints: context.callPoints,
+    services: context.services,
   });
 
   if (invalidBytes !== null) {
     return invalidBytes;
   }
 
-  const artifactExists = await services.artifacts.hasArtifact(
-    vertex.libraryName
-  );
+  const artifactExists = isArtifact(vertex.artifact);
 
   if (!artifactExists) {
     return buildValidationError(
       vertex,
-      `Library with name '${vertex.libraryName}' doesn't exist`,
-      callPoints
+      `Artifact not provided for library '${vertex.label}'`,
+      context.callPoints
     );
   }
 
-  const artifact = await services.artifacts.getArtifact(vertex.libraryName);
   const argsLength = vertex.args.length;
 
-  const iface = new ethers.utils.Interface(artifact.abi);
+  const iface = new ethers.utils.Interface(vertex.artifact.abi);
   const expectedArgsLength = iface.deploy.inputs.length;
 
   if (argsLength !== expectedArgsLength) {
     return buildValidationError(
       vertex,
-      `The constructor of the library '${vertex.libraryName}' expects ${expectedArgsLength} arguments but ${argsLength} were given`,
-      callPoints
+      `The constructor of the library '${vertex.label}' expects ${expectedArgsLength} arguments but ${argsLength} were given`,
+      context.callPoints
     );
   }
