+++ packages/core/src/internal/validation/dispatch/validateCall.ts
@@ -1,13 +1,13 @@
 import { ethers, BigNumber } from "ethers";
 
-import { CallDeploymentVertex } from "../../internal/types/deploymentGraph";
-import { VertexResultEnum } from "../../internal/types/graph";
-import { isParameter } from "../../internal/utils/guards";
+import { CallDeploymentVertex } from "../../types/deploymentGraph";
+import { VertexResultEnum } from "../../types/graph";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
   ValidationVertexVisitResult,
 } from "../../types/validation";
+import { isParameter } from "../../utils/guards";
 
 import {
   buildValidationError,
