+++ packages/core/src/internal/validation/dispatch/validationDispatch.ts
@@ -1,4 +1,4 @@
-import { DeploymentGraphVertex } from "../../internal/types/deploymentGraph";
+import { DeploymentGraphVertex } from "../../types/deploymentGraph";
 import {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
